server/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
firebase-admin==6.5.0
flask-cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.1.3
flasgger==0.9.7.1
